# Avoid large objects allocations and reuse everthing