Main part of this [PR](https://github.com/reactjs/React.NET/pull/522)
I propose to mark the old methods that do not take the TextWriter as obsolete, to avoid errors if it is overridden
@dustinsoftware @Daniel15