@dustinsoftware 2018-04-23
Ok! I got around to testing this and it looks good. However, I'd like to get some more concrete benchmarks documented before merging this in. On my workstation there was no noticeable speed or GC improvements with this change, but I'm unsure if the component and props are so small with my sample that it doesn't show.

[Here the tool I put together, based on BenchmarkDotNet](https://github.com/dustinsoftware/React.NET/blob/benchmark/tests/React.Tests.Benchmarks/Program.cs)... feel free to suggest a better code sample I can use to show the benefits of this change. I'll do some more testing and see what I can find out. My goal is to add this benchmark tool to this project so we can find other improvements in the render pipeline.