@dustinsoftware 2018-04-23
Before this change:

``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.16299.371 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i5-4570 CPU 3.20GHz (Haswell), 1 CPU, 4 logical and 4 physical cores
Frequency=3117782 Hz, Resolution=320.7408 ns, Timer=TSC
  [Host]     : .NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  DefaultJob : .NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0


```
|                      Method |     Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Allocated |
|---------------------------- |---------:|----------:|----------:|---------:|--------:|----------:|
|  HtmlHelperExtensions_React | 50.56 ms | 0.4835 ms | 0.4286 ms | 250.0000 | 62.5000 |  13.22 MB |
| Environment_CreateComponent | 47.62 ms | 0.4689 ms | 0.4386 ms | 250.0000 | 62.5000 |   9.78 MB |


After:

``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.16299.371 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i5-4570 CPU 3.20GHz (Haswell), 1 CPU, 4 logical and 4 physical cores
Frequency=3117782 Hz, Resolution=320.7408 ns, Timer=TSC
  [Host]     : .NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  DefaultJob : .NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0


```
|                      Method |     Mean |     Error |    StdDev |    Gen 0 |    Gen 1 | Allocated |
|---------------------------- |---------:|----------:|----------:|---------:|---------:|----------:|
|  HtmlHelperExtensions_React | 46.21 ms | 0.6098 ms | 0.5704 ms | 250.0000 | 125.0000 |   7.46 MB |
| Environment_CreateComponent | 46.74 ms | 0.5432 ms | 0.5081 ms | 250.0000 | 125.0000 |   9.18 MB |